refactor : adding interface for ProfileService and implentation
